Fixed missing state in storybook
Also moved the default challenge to be in examples
