+++ src/examples/default-challenge.ts
@@ -1,0 +1,26 @@
+import { Challenge } from '../model/ChallengeConfiguration'
+
+const defaultChallenge: Challenge = {
+  start: {
+    firstPoint: {
+      lon: -122.4,
+      lat: 37.7,
+    },
+    secondPoint: {
+      lon: -122.4,
+      lat: 37.8,
+    },
+  },
+  finish: {
+    firstPoint: {
+      lon: -122.5,
+      lat: 37.7,
+    },
+    secondPoint: {
+      lon: -122.5,
+      lat: 37.8,
+    },
+  },
+}
+
+export { defaultChallenge }
